changed default likelihood precision to double

+++ src/bayes.h
@@ -85,7 +85,7 @@ typedef long RandLong;
 #define MRBFLT_NEG_MAX (-DBL_MAX)  /* maximum possible negative value that can be stored in MrBFlt */
 typedef double MrBFlt;      /* double used for parameter values and generally for floating point values, 
                                if set to float MPI would not work becouse of use MPI_DOUBLE */
-typedef float CLFlt;        /* single-precision float used for cond likes (CLFlt) to increase speed and reduce memory requirement */
+typedef double CLFlt;        /* single-precision float used for cond likes (CLFlt) to increase speed and reduce memory requirement */
                             /* set CLFlt to double if you want increased precision */
                             /* NOTE: CLFlt = double not compatible with SSE_ENABLED */
